refactor(api): keep newChartFile.ts file and mount it to /chartFile route
