+++ route/src/api/app.ts
@@ -3,6 +3,7 @@ import { cors } from "hono/cors";
 import briefApp from "./brief.js";
 import { backendOrigin, Bindings, immutable } from "../env.js";
 import chartFileApp from "./chartFile.js";
+import newChartFileApp from "./newChartFile.js";
 import playFileApp from "./playFile.js";
 import seqFileApp from "./seqFile.js";
 import seqPreviewApp from "./seqPreview.js";
@@ -101,6 +102,10 @@ const apiApp = async (config: {
     .route(
       "/chartFile",
       await chartFileApp({ getConnInfo: config.getConnInfo })
+    )
+    .route(
+      "/chartFile",
+      await newChartFileApp({ getConnInfo: config.getConnInfo })
     )
     .all("/newChartFile", (c) =>
       c.redirect(new URL("/api/chartFile", backendOrigin(c)), 307)
