+++ route/src/api/chartFile.ts
@@ -85,189 +85,6 @@ const chartFileApp = async (config: {
     Bindings: Bindings;
     Variables: ChartFileAppVars;
   }>({ strict: false })
-    .post(
-      "/",
-      describeRoute({
-        description:
-          "Create a new chart. " +
-          "The chart data should be in MessagePack format, and " +
-          `must be the latest format (Chart17) or one version earlier (Chart15). ` +
-          `The chart data may be compressed using ${supportedEncodings.join(", ")} (in that case Content-Encoding header must be set.) ` +
-          "Returns the chart ID (cid) of the newly created chart. " +
-          `This endpoint is rate limited to one request per ${rateLimit.newChartFile / 60} minutes. `,
-        requestBody: {
-          description:
-            "Chart data in MessagePack format. See also response type of GET /api/chartFile.",
-          required: true,
-          content: {
-            "application/vnd.msgpack": {
-              schema: { anyOf: [docRefs("Chart17"), docRefs("Chart15")] },
-            },
-          },
-        },
-        parameters: [
-          {
-            name: "Content-Encoding",
-            in: "header",
-            description: "Encoding applied to the request body",
-            schema: { type: "string" },
-          },
-        ],
-        responses: {
-          200: {
-            description: "Successful response",
-            content: {
-              "application/json": {
-                schema: resolver(v.object({ cid: CidSchema() })),
-              },
-            },
-            headers: {
-              ...etagHeaderDoc,
-            },
-          },
-          400: {
-            description: "password not specified in the chart data",
-            content: {
-              "application/json": {
-                schema: resolver(await errorLiteral("noPasswd")),
-              },
-            },
-          },
-          409: {
-            description: `chart version is older than ${currentChartVer - 1}`,
-            content: {
-              "application/json": {
-                schema: resolver(await errorLiteral("oldChartVersion")),
-              },
-            },
-          },
-          413: {
-            description: "Chart file too large",
-            content: {
-              "application/json": {
-                schema: resolver(
-                  await errorLiteral("tooLargeFile", "tooManyEvent")
-                ),
-              },
-            },
-          },
-          415: {
-            description:
-              "Invalid chart format, or given Content-Encoding is unsupported",
-            content: {
-              "application/json": {
-                schema: resolver(
-                  v.union([
-                    await validationErrorSchema("invalidChart"),
-                    await errorLiteral(
-                      "invalidChart",
-                      "unsupportedContentEncoding",
-                      "invalidContentEncoding"
-                    ),
-                  ])
-                ),
-              },
-            },
-            headers: {
-              "Accept-Encoding": {
-                description: `Supported encoding type (${supportedEncodings.join(", ")})`,
-                schema: { type: "string" },
-              },
-            },
-          },
-          429: {
-            description: "Rate limited",
-            content: {
-              "application/json": {
-                schema: resolver(await errorLiteral("tooManyRequest")),
-              },
-            },
-            headers: {
-              "Retry-After": {
-                description: "Number of seconds to wait before retrying",
-                schema: { type: "integer" },
-              },
-            },
-          },
-        },
-      }),
-      async (c) => {
-        const ip = getIp(c, config.getConnInfo);
-        const chartBuf = await c.req.arrayBuffer();
-        const pSecretSalt = secretSalt(env(c));
-        const db = await c.get("db")();
-
-        if (!(await updateIp(env(c), db, ip, "newChartFile"))) {
-          return c.json(
-            {
-              message: "tooManyRequest",
-            },
-            429,
-            { "retry-after": rateLimit.newChartFile.toString() }
-          );
-        }
-
-        let newChart: Chart15 | Chart17;
-        try {
-          newChart = msgpack.decode(chartBuf) as Chart15 | Chart17;
-          if (newChart.ver < currentChartVer - 1) {
-            return c.json({ message: "oldChartVersion" }, 409);
-          }
-          newChart = validateChartWithoutConvert(newChart) as Chart15 | Chart17;
-        } catch (e) {
-          throw new HTTPException(415, { message: "invalidChart", cause: e });
-        }
-
-        if (numEvents(newChart) > chartMaxEvent) {
-          throw new HTTPException(413, {
-            message: "tooManyEvent",
-          });
-        }
-
-        // update Time
-        const updatedAt = new Date().getTime();
-
-        let cid: string;
-        while (true) {
-          cid = Math.floor(Math.random() * 890000 + 110000).toString();
-          if (
-            await db
-              .collection<ChartEntryCompressed>("chart")
-              .countDocuments({ cid }, { limit: 1 })
-          ) {
-            continue;
-          } else {
-            break;
-          }
-        }
-
-        await db
-          .collection<ChartEntryCompressed>("chart")
-          .insertOne(
-            await zipEntry(
-              await chartToEntry(
-                newChart,
-                cid,
-                updatedAt,
-                ip,
-                await getYTDataEntry(
-                  c.var.logger,
-                  env(c),
-                  db,
-                  newChart.ytId
-                ).catch(() => undefined),
-                pSecretSalt,
-                null
-              )
-            )
-          );
-
-        const newEntry = await getChartEntryCompressed(db, cid, null);
-        return c.json({ cid: cid }, 200, {
-          "ETag": await calcETag(newEntry),
-        });
-      }
-    )
     .on(
       ["GET", "PUT", "POST", "DELETE"],
       "/:cid",
